preserve the link order and account for future benefits
